Switch Cloudflare deploy to Workers Static Assets model
Cloudflare's Git integration creates a Worker (running `wrangler deploy`), not a
Pages project. The previous Pages-style wrangler.jsonc (pages_build_output_dir +
functions/) made that build fail: "you have run `wrangler deploy` on a Pages
project … Missing entry-point to Worker script or to assets directory."

Convert to the Workers Static Assets model so `wrangler deploy` succeeds:

- wrangler.jsonc: `main` + `assets` (directory ./dist, binding ASSETS,
  run_worker_first ["/"]). `name` set to "word-to-markdown-js" to match the
  Cloudflare project created from the repo.
- worker/index.js: replaces functions/index.js. Default fetch handler does the
  Accept-Language redirect on "/" (preserving the query string, cookie-once) and
  falls through to env.ASSETS.fetch() for everything else. Pure pickLocale() is
  unchanged and still unit-tested.
- Remove functions/index.js (Pages-only convention).
- Docs/comments updated for the Workers flow (`wrangler deploy`, *.workers.dev,
  custom domain under the Worker, `npx wrangler dev` for local checks).

Still a static Astro build (output: 'static'); inert on GitHub Pages (serves
only ./dist). Integration-tested locally via `wrangler dev`: 302-by-language,
no redirect for English / when cookied, /de/ served from assets, query string
preserved, English root falls through to the real index.html.

Verified: 137 unit + 19 e2e pass, lint clean, full build passes all checks.
