+++ web/layouts/Layout.astro
@@ -232,8 +232,8 @@ const footerLinks = [
       </p>
     </footer>
     <script is:inline>
-      // Record the current language so the edge redirect (Cloudflare Pages,
-      // functions/index.js) won't auto-redirect on future visits — the language
+      // Record the current language so the edge redirect (Cloudflare Worker,
+      // worker/index.js) won't auto-redirect on future visits — the language
       // switcher stays in control. Harmless on GitHub Pages: it sets a cookie
       // that nothing reads there.
       try {
